pyproject.toml
[project]
name = "s4lt"
version = "0.8.0"
description = "Sims 4 Linux Toolkit - Manage mods on Linux and Steam Deck"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "S4LT Contributors"}
]
keywords = ["sims4", "mods", "linux", "steamdeck", "modding"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment :: Simulation",
]

dependencies = [
    # CLI
    "click>=8.0",
    "rich>=13.0",

    # Web UI
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",

    # Desktop App
    "pywebview>=5.0",
    "pystray>=0.19",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "httpx>=0.26.0",
    "pyinstaller>=6.0",
    "ruff>=0.1.0",
]

[project.scripts]
s4lt = "s4lt.cli.main:cli"
s4lt-desktop = "s4lt.desktop.launcher:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["s4lt*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
